fixed the wallet module

+++ src/server/modules/wallet/wallet.route.ts
@@ -12,14 +12,14 @@ export const walletRoute = new Hono()
   .post("/", authGuard, jsonValidator(walletValidation.createWallet), async (ctx) => {
     const dto = ctx.req.valid("json");
     const user = ctx.get("user");
-    await walletService.createWalletWithInitialTransaction(dto, user._id);
+    await walletService.createWalletWithInitialTransaction({ ...dto, ownerId: user._id });
     return ctx.json(ResponseDto.success("Wallet created successfully"));
   })
   // Get Wallets
   .get("/", authGuard, queryValidator(walletValidation.getAllWallets), async (ctx) => {
     const query = ctx.req.valid("query");
     const user = ctx.get("user");
-    const { wallets, meta } = await walletService.getWallets(query, user._id);
+    const { wallets, meta } = await walletService.getWallets({ query, userId: user._id });
     return ctx.json(ResponseDto.success({ message: "Wallets fetched successfully", meta, data: wallets }));
   })
   // Update Wallet
@@ -27,21 +27,21 @@ export const walletRoute = new Hono()
     const dto = ctx.req.valid("json");
     const user = ctx.get("user");
     const id = ctx.req.param("id");
-    await walletService.updateWallet(dto, id, user._id);
+    await walletService.updateWallet({ id, dto, userId: user._id });
     return ctx.json(ResponseDto.success("Wallet updated successfully"));
   })
   // Delete Wallet
   .delete("/:id", authGuard, async (ctx) => {
     const user = ctx.get("user");
     const id = ctx.req.param("id");
-    await walletService.deleteWallet(id, user._id);
+    await walletService.deleteWallet({ id, userId: user._id });
     return ctx.json(ResponseDto.success("Wallet deleted successfully"));
   })
   // Transfer Money
   .post("/transfer", authGuard, jsonValidator(walletValidation.transfer), async (ctx) => {
     const dto = ctx.req.valid("json");
     const user = ctx.get("user");
-    await walletService.walletTransfer(dto, user._id);
+    await walletService.walletTransfer({ dto, userId: user._id });
     return ctx.json(ResponseDto.success("Wallet transfer has been successful"));
   });
